Added getVertices() method to GeoEntity and implemented for Line, Mesh, Polygon and Feature. Added toArray method for Vertex with dimensions and rounding. Fixed WKT conversion methods.

+++ src/model/GeoEntity.js
@@ -6,7 +6,7 @@ define([
   'atlas/model/Style',
   'atlas/util/DeveloperError',
   'atlas/util/mixin'
-], function (Event, EventTarget, Colour, Style, DeveloperError, mixin) {
+], function(Event, EventTarget, Colour, Style, DeveloperError, mixin) {
 
   /**
    * @typedef atlas.model.GeoEntity
@@ -38,7 +38,7 @@ define([
    * @extends atlas.events.EventTarget
    * @class atlas.model.GeoEntity
    */
-   GeoEntity = EventTarget.extend(/** @lends atlas.model.GeoEntity# */ {
+  GeoEntity = EventTarget.extend(/** @lends atlas.model.GeoEntity# */ {
     /**
      * The ID of the GeoEntity
      * @type {String}
@@ -151,7 +151,7 @@ define([
      */
     _selected: false,
 
-    _init: function (id, args) {
+    _init: function(id, args) {
       if (typeof id === 'object') {
         args = id;
         id = args.id;
@@ -196,9 +196,19 @@ define([
       throw new DeveloperError('Can not call abstract method "getCentroid" of GeoEntity');
     },
 
+    /**
+     * @returns {Array.<atlas.model.Vertex>}
+     * @abstract
+     */
+    getVertices: function() {
+      throw new DeveloperError('Can not call abstract method "getVertices" of GeoEntity');
+    },
+
     getChildren: function() {
       // TODO(bpstudds): Adding children and removing children needs support.
-      if (this._children !== null) { return this._children; }
+      if (this._children !== null) {
+        return this._children;
+      }
       this._children = this._entityManager.getByIds(this._childrenIds);
       return this._children;
     },
@@ -214,7 +224,7 @@ define([
     /**
      * @returns {[atlas.model.Handle]} An array of Handles used to edit the GeoEntity.
      */
-    getEditingHandles: function () {
+    getEditingHandles: function() {
       return [];
     },
 
@@ -251,7 +261,9 @@ define([
      * <code>component</code> is not given, the GeoEntity as a whole.
      */
     isDirty: function(component) {
-      if (component === undefined) { return Object.keys(this._dirty).length > 0; }
+      if (component === undefined) {
+        return Object.keys(this._dirty).length > 0;
+      }
       return component in this._dirty;
     },
 
@@ -269,7 +281,9 @@ define([
      * @returns {atlas.model.Style} The old style, or null if it was not changed.
      */
     setStyle: function(style) {
-      if (this._style === style) { return null; }
+      if (this._style === style) {
+        return null;
+      }
       this.setDirty('style');
 
       // Only change style if the new style is different so _previousStyle isn't clobbered.
@@ -281,14 +295,14 @@ define([
     /**
      * @returns {atlas.model.Style} The style of the GeoEntity.
      */
-    getStyle: function () {
+    getStyle: function() {
       return this._style;
     },
 
     /**
      * @returns {Boolean} Whether the GeoEntity is currently renderable.
      */
-    isRenderable: function () {
+    isRenderable: function() {
       return Object.keys(this._dirty).length === 0;
     },
 
@@ -299,7 +313,7 @@ define([
      * @returns {Object} The geometry data.
      */
     getGeometry: function() {
-        return this._geometry;
+      return this._geometry;
     },
 
     /**
@@ -309,7 +323,7 @@ define([
      * @returns {Object} The appearance data.
      */
     getAppearance: function() {
-        return this._appearance;
+      return this._appearance;
     },
 
     // -------------------------------------------
@@ -324,8 +338,10 @@ define([
      * @param {Number} [newStyle.borderWidth] - The new border width colour.
      * @returns {Object} A mapping of parameters that have been changed to their old value.
      */
-    modifyStyle: function (newStyle) {
-      if (Object.keys(newStyle).length <= 0) { return {}; }
+    modifyStyle: function(newStyle) {
+      if (Object.keys(newStyle).length <= 0) {
+        return {};
+      }
 
       this.setDirty('style');
       var oldStyle = {};
@@ -384,7 +400,7 @@ define([
      *
      * @abstract
      */
-    rotate: function (rotation) {
+    rotate: function(rotation) {
       throw new DeveloperError('Can not call abstract method "rotate" of GeoEntity');
     },
 
@@ -401,13 +417,15 @@ define([
      * be overridden on subclasses to accomplish any cleanup that
      * may be required.
      */
-    remove: function () {
-      if (!this._eventManager) { return; }
-      this._entityManager.remove(this._id);
-      this._eventManager.dispatchEvent(new Event(new EventTarget(),
-        'entity/remove', {
-          id: this.getId()
-        }));
+    remove: function() {
+      // TODO(aramk) We should try to keep consistent with these - either all entities have
+      // references to managers or none do - otherwise we could have discrepancies in the entity
+      // manger like a removed entity still being referenced.
+      this._entityManager && this._entityManager.remove(this._id);
+      this._eventManager && this._eventManager.dispatchEvent(new Event(new EventTarget(),
+          'entity/remove', {
+            id: this.getId()
+          }));
     },
 
     /**
@@ -415,7 +433,7 @@ define([
      * @returns {Boolean} Whether the GeoEntity is shown.
      * @abstract
      */
-    show: function () {
+    show: function() {
       throw new DeveloperError('Can not call abstract method "show" of GeoEntity');
     },
 
@@ -424,14 +442,14 @@ define([
      * @returns {Boolean} Whether the GeoEntity is hidden.
      * @abstract
      */
-    hide: function () {
+    hide: function() {
       throw new DeveloperError('Can not call abstract method "hide" of GeoEntity');
     },
 
     /**
      * Toggles the visibility of the GeoEntity.
      */
-    toggleVisibility: function () {
+    toggleVisibility: function() {
       if (this.isVisible()) {
         this.hide();
       } else {
@@ -447,25 +465,27 @@ define([
      * Handles the GeoEntities behaviour when it is selected.
      * @abstract
      */
-    onSelect: function () {},
+    onSelect: function() {
+    },
 
     /**
      * Handles the GeoEntities behaviour when it is deselected.
      * @abstract
      */
-    onDeselect: function () {},
+    onDeselect: function() {
+    },
 
     /**
      * Enables 'editing' of the GeoEntity using keyboard input.
      */
-    onEnableEditing: function () {
+    onEnableEditing: function() {
       return;
       // TODO(bpstudds): Move this functionality to an EditManager module.
       this._editEventHandler = this._eventManager.addEventHandler('intern', 'input/keydown',
-          function (args) {
+          function(args) {
             // TODO(bpstudds): Replace 'magic numbers' with constants. Probably should update keycode.js library for this.
-            if (!args.modifiers.shiftKey && !args.modifiers.metaKey &&
-                !args.modifiers.altKey && !args.modifiers.ctrlKey) {
+            if (!args.modifiers.shiftKey && !args.modifiers.metaKey && !args.modifiers.altKey &&
+                !args.modifiers.ctrlKey) {
               switch (args.key) {
                 case 95: // underscore/minus beside backspace key
                   this.scale({x: 0.95, y: 0.95, z: 0.95});
@@ -474,10 +494,10 @@ define([
                   this.scale({x: 1.05, y: 1.05, z: 1.05});
                   break;
                 case 37: // left
-                  this.rotate({x: 0, y: 0, z:5});
+                  this.rotate({x: 0, y: 0, z: 5});
                   break;
                 case 39: // right
-                  this.rotate({x: 0, y: 0, z:-5});
+                  this.rotate({x: 0, y: 0, z: -5});
                   break;
               }
             }
@@ -492,21 +512,20 @@ define([
     /**
      * Disables editing of the GeoEntity.
      */
-    onDisableEditing: function () {
+    onDisableEditing: function() {
       return;
       this._editEventHandler && this._editEventHandler.cancel();
       this._editingHandles.forEach(function(handle) {
         handle.remove();
       })
     }
-    
+
   });
 
   // -------------------------------------------------
   // Statics
   // -------------------------------------------------
   GeoEntity.DEFAULT_STYLE = new Style(Colour.GREEN, Colour.GREEN, 5)
-
 
   return GeoEntity;
 });
